fix(normalization): set content-length to 0 and remove content-type for put/patch requests without a body
The GitHub API has several endpoints that require a PUT or a PATCH verb
with an empty body. In that case it’s unclear what content-type to set,
or if a content-type header is to be set at all. So we remove it from our
fixtures in that case but set content-length to 0 as required via:
https://developer.github.com/v3/#http-verbs